+++ packages/router/src/createRouter.ts
@@ -112,10 +112,10 @@ export default function createRouter<O = HistoryOptions>(
     emitImmediate(response, navigation);
   }
 
-  const { emitRedirects = true } = options;
+  const { invisibleRedirects = false } = options;
 
   function emitImmediate(response: Response, navigation: Navigation) {
-    if (!response.redirect || emitRedirects) {
+    if (!response.redirect || !invisibleRedirects) {
       mostRecent.response = response;
       mostRecent.navigation = navigation;