+++ packages/types/src/index.ts
@@ -34,7 +34,7 @@ export interface PathOptions {
 export interface RouterOptions<O = HistoryOptions> {
   route?: Array<Interaction>;
   sideEffects?: Array<Observer>;
-  emitRedirects?: boolean;
+  invisibleRedirects?: boolean;
   external?: any;
   history?: O;
 }